WIP CLI safety flags and other cli settings

+++ RequestHandlers/OrgHandlers.cs
@@ -1,5 +1,6 @@
 using tankman.Services;
 using tankman.Http;
+using tankman.Utils;
 
 namespace tankman.RequestHandlers;
 
@@ -31,8 +32,11 @@ public static class OrgHandlers
   }
 
 
-  public static async Task<IResult> DeleteOrgAsync(string orgId)
+  public static async Task<IResult> DeleteOrgAsync(string orgId, string safetyKey)
   {
+    if (Settings.SafetyKey != null && safetyKey != Settings.SafetyKey)  {
+      return TypedResults.BadRequest("Missing org deletion key.");
+    }
     return ApiResult.ToResult(await OrgService.DeleteOrgAsync(orgId));
   }
 }
